refactor: Moved book services folder to its parent folder
