+++ src/modules/book/book.module.ts
@@ -3,8 +3,8 @@ import { BookService } from './book.service';
 import { BookController } from './book.controller';
 import { Book } from './entities/book.entity';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { CreateBookService } from './dto/services/create-book.service';
-import { FindAllBooksService } from './dto/services/find-all-books.service';
+import { CreateBookService } from './services/create-book.service';
+import { FindAllBooksService } from './services/find-all-books.service';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Book])],
