+++ src/modules/book/book.service.spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { BookService } from './book.service';
-import { CreateBookService } from './dto/services/create-book.service';
-import { FindAllBooksService } from './dto/services/find-all-books.service';
+import { CreateBookService } from './services/create-book.service';
+import { FindAllBooksService } from './services/find-all-books.service';
 
 const mockBookRepository = {
   // You can add mock implementations for any methods you'll be using.
